\name{mlfilms-package}
\alias{mlfilms-package}
\alias{mlfilms}
\docType{package}
\title{
Calculation of the optical response of multilayer thin films

}
\description{
Calculates the optical response of multilayer thin films using a transfer matrix method.
}
\details{
\tabular{ll}{
Package: \tab mlfilms\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-04-04\cr
License: \tab GPL\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Tom Constant

Maintainer: Tom Constant <t.j.constant@gmail.com>

}
\references{
Introduction to Optics. Pedrotti, Pedrotti and Pedrotti, 3rd Edition. pp. 476 onwards.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
