% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength_scan.R
\name{wavelength_scan}
\alias{wavelength_scan}
\title{Spectra calculation at a fixed angle}
\usage{
wavelength_scan(layers, wavelengths = seq(3.5e-07, 8.5e-07, length.out = 500),
  angle = 0, polarisation = "p", incident_medium.index = complex(real = 1,
  imaginary = 0), exit_medium.index = complex(real = 1, imaginary = 0),
  dispersive.function = "none", dispersive.layers = NA, show.progress = F)
}
\arguments{
\item{layers}{A list object containing the stack parameters. Must include index, thickness and repetitions. See details and examples for more information.}

\item{wavelengths}{The wavelength range of the calculated spectra, in meters. The default covers the visible range from 350 nm to 850 nm.}

\item{angle}{Fixed angle in degrees. Default is 0.}

\item{polarisation}{Linear polarisation of the light. Acceptable arguments are 'p' (Transverse Magnetic) or 's' (Transverse Electric).}

\item{incident_medium.index}{The global incident medium. Default is n=1+0i (air)}

\item{exit_medium.index}{The global exit medium. Default is n=1+0i (air)}

\item{dispersive.function}{For dispersive materials only, specify a function which returns the refractive index as a function of wavlength for the layer defined using the dispersive.layers varible.}

\item{dispersive.layers}{Vector of layers to replace with dispersive.function. Starting with the top layer in the multilayer stack = 1.}
}
\value{
Returns a dataframe with the wavelength, Refelctivity, Transmission and Absorbtion
}
\description{
Function to calculate the reflectivity, transmission and absorption as a function of wavelength for a given multilayer film.
}
\details{
The layers list should be constructed like so:

\code{layers <- list(index = ..., thickness = ..., repetitions = ...)}

where index and thickness are vectors containing the stack parameters in order from the top interface to the bottom. Repetitions is an integer repeating the stack.
}
\examples{

layers<-list(index=c(2.35+0i,1.38+0i),thickness=c(550e-9/(4*2.35),550e-9/(4*1.38)),repetitions=6)

R_highlowStack6<-wavelength_scan(incident_medium.index=1+0i,exit_medium.index=1.52+0i,layers=layers)

plot(R_highlowStack6$wavelength,R_highlowStack6$Reflection,type='l',lwd=2, ylim=c(0,1))
title("H/L index stack (N=6 & 2): Pedrotti Figure 22-9")
}
\references{
Introduction to Optics 3rd Edition, Pearson international edition by Frank L. Pedrotti, Leno Matthew Pedrotti, Leno S. Pedrotti
}
